requirements.txt
streamlit
sentence-transformers
pinecone
google-generativeai
python-dotenv
plotly
pandas
arxiv
tqdm
streamlit-lottie